+++ docs/changelog.md
@@ -9,3 +9,6 @@
   - added `S3FileInfo`
   - added `S3Filesystem`
   - added `S3FilesystemContents`
+* Added internal helpers
+  - added `CallListObjectsV2'
+  - added `GetFilesystemContents`